New room with slightly new roles
Voyagers are now being spawned in the new room to harvest from the room directly above. These voyagers are vital to our success because there are not enough resources in my new room to sustain the village like it must. Two towers have already been put in place for this new room as well. We are keeping the voyager's body components at a cost of 250 while we bumped the harvesters body components to 650. We are hoping to increase both of these soon but we must not exhaust our resources.